Finished database design assignment
Added names to top of convert.py
Added third query to queries.sql and added more details to the ones
already there.
Made the database dump in database-dump.sql
Added extra file not part of the assignment named copy-databases.sql
that was useful when copying over the databases after creating.